requirement.txt
absl-py==1.4.0
astor==0.8.1
audioread==3.0.1
cached-property==1.5.2
certifi==2020.6.20
cffi==1.15.1
click==8.0.4
dataclasses==0.8
decorator==5.1.1
editdistance==0.6.2
gast==0.5.4
grpcio==1.48.2
h5py==3.1.0
importlib-metadata==4.8.3
joblib==1.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.3.7
mock==5.1.0
numba==0.48.0
numpy==1.19.5
packaging==21.3
Pillow==8.4.0
pip==21.3.1
protobuf==4.21.0
pycparser==2.21
pyparsing==3.1.1
PyYAML==6.0.1
resampy==0.3.1
scikit-learn==0.24.2
scipy==1.5.4
setuptools==36.4.0
six==1.16.0
SoundFile==0.10.2
svgwrite==1.4.3
tensorboard==1.13.1
tensorboardX==2.6.2.2
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
threadpoolctl==3.1.0
torch==1.0.1
torchvision==0.2.2
tqdm==4.19.9
Tree==0.2.4
typing_extensions==4.1.1
Werkzeug==2.0.3
wheel==0.37.1
zipp==3.6.0
